dom/tests/mochitest/geolocation/test_geoGetCurrentPositionBlockedInInsecureContext.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1269531
-->
<head>
  <title>Test for Bug 1269531</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1269531">Mozilla Bug 1269531</a>
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>
  <pre id="test">
  <script class="testbody" type="text/javascript">

  SimpleTest.waitForExplicitFinish();

  // The test succeeds if the error callback is called because we expect
  // failure when a getCurrentPosition() request is submitted in a non
  // secure context.
  function successCallback(position) {
    ok(false, "Success callback is not expected to be called");
    SimpleTest.finish();
  }

  function errorCallback(error) {
    ok(true, "Check for the error callback to be called for insecure requests");
    is(error.code, error.PERMISSION_DENIED, "Check error code for insecure requests");
    SimpleTest.finish();
  }

  // Insecure requests should be blocked, for that we enable the relevant pref.
  SpecialPowers.pushPrefEnv({"set": [["geo.security.allowinsecure", false]]}, function() {
    force_prompt(true, function() {
      navigator.geolocation.getCurrentPosition(successCallback, errorCallback);
    })
  });

  </script>
  </pre>
</body>
</html>
